Seperate Linux build code to non-linux
